20160406/2-EstructuraMavenModelClasses/src/main/java/com/indra/formacio/model/Employee.java
package com.indra.formacio.model;

import java.text.SimpleDateFormat;
import java.util.Date;

/**
 * Objects of this class represent an Employee
 * 
 * @author ealcalal
 *
 */
public class Employee {
	
	SimpleDateFormat	sdf = new SimpleDateFormat("dd/mm/yyyy");	
	
	protected Long id;
	protected String name;
	protected String surname;
	protected Date birthday;
	protected Float percentCustomers;
	protected Date percentDate;
	
//	public Employee (long id, String name, String surname, String birthday){
//		this.id = id;
//		this.name = name;
//		this.surname = surname;
//		//Alerta con el SURROUND!!
//		try {
//			this.birthday = sdf.parse(birthday);
//		} catch (ParseException e) {
//			System.out.println("Error seteando la fecha!!");
//			e.printStackTrace();
//			this.birthday = new Date();
//		}
//	}
	
	public Long getId() {
		return id;
	}
	public void setId(Long id) {
		this.id = id;
	}
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getSurname() {
		return surname;
	}
	public void setSurname(String surname) {
		this.surname = surname;
	}
	public Date getBirthday() {
		return birthday;
	}
	public void setBirthday(Date birthday) {
		this.birthday = birthday;
	}
	public Float getPercentCustomers() {
		return percentCustomers;
	}
	public void setPercentCustomers(Float percentCustomers) {
		this.percentCustomers = percentCustomers;
	}
	public Date getPercentDate() {
		return percentDate;
	}
	public void setPercentDate(Date percentDate) {
		this.percentDate = percentDate;
	}
	
}
